Fix tests Numpy 2.0.0 NEP 51 representation strs
Numpy 2.0.0 adopted NEP 51, meaning the string representation of NumPy
data types changed to e.g. `np.float64(10.0)` to offer differentiation
against Python's internal data types.

This caused the tests to fail as all the tests are hardcoded against
the old style.

This commit introduces a quick fix, which is to ask numpy to print with
the old style.

https://numpy.org/devdocs/release/2.0.0-notes.html#representation-of-numpy-scalars-changed

https://numpy.org/neps/nep-0051-scalar-representation.html#nep51